refactor: 강의 카드 여러 파일로 리팩터 #49
