--- src/components/lecture/BookmarkButton.tsx
@@ -1,5 +1,0 @@
-const BookmarkButton = () => {
-  return <div></div>
-}
-
-export default BookmarkButton
